Put RepFR back to the way it was supposed to be and minor improvements

+++ Assets/Scripts/FRExperiment.cs
@@ -123,7 +123,7 @@ public class FRExperiment : ExperimentBase {
   }
 
   protected void Rest(StateMachine state) {
-    int duration = (int)Config.restDuration;
+    int duration = Config.restDuration;
     state.IncrementState();
     manager.Do(new EventBase<string, string>(manager.ShowText, "orientation stimulus", "+"));
     ReportEvent("rest", null);
